app/requirements.txt
click
Django
django-ckeditor
django-filter
django-js-asset
django-solo
djangorestframework
djangorestframework-datatables
itsdangerous
macholib
MarkupSafe
numpy
pandas
Pillow
python-dateutil
pytz
six
sqlparse
Werkzeug
whichcraft
whitenoise
xlrd
reportlab
xhtml2pdf